scheduler moved to app layer
