+++ packages/api/src/application/notifications/notifications.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { NotificationsServiceInterface } from 'src/domain/notifications/notifications-service.abstract';
 import { NotificationsSender, NotificationsSenderToken } from '../../domain/notifications/notifications-sender.interface';
-import { MailModule } from '../../infrastructure/notifications/mail/mail.module';
-import { MailSender } from '../../infrastructure/notifications/mail/services/mail-sender.service';
+import { MailModule } from '../../infrastructure/mail/mail.module';
+import { MailSender } from '../../infrastructure/mail/services/mail-sender.service';
 import { NotificationsService } from './services/notifications.service';
 
 @Module({
