+++ packages/api/src/application/weather-scheduler/types/weather-update.options.ts
@@ -1,5 +1,5 @@
 import { Frequency } from 'src/domain/subscription/frequency.vo';
-import { NotificationSubjects } from '../../../../application/notifications/constants/notification-subjects.enum';
+import { NotificationSubjects } from '../../notifications/constants/notification-subjects.enum';
 
 export type WeatherUpdateOptions = {
   frequency: Frequency;
