+++ packages/api/src/infrastructure/mail/constants/template.type.ts
