+++ packages/api/src/infrastructure/mail/constants/templates.enum.ts
