+++ packages/api/src/infrastructure/mail/mail.module.ts
