+++ packages/api/src/infrastructure/mail/services/mail-template.service.ts
